docs: 💄 update landing page layout (#362)
# Description

A small update of the landing page layout, including simplification of
some of the text.

I want to do more to improve this (including the brand), but I think
that this is a nice improvement for now 💅 🎨
 
Let me know what you think.

Closes #361 

This PR needs a an in-depth review.

## Checklist

- [X] Formatted Markdown
- [X] Ran `just run-all`